app/requirements.txt
requests==2.25.1
google-cloud-aiplatform==1.22.0
tensorflow==2.9.0
streamlit== 1.3.1
google_api_python_client==1.12.8
protobuf==3.20.3
torch==1.12.1
torchvision==0.13.1
altair==4.0

